refactor: move Reports views into Views directory
