+++ give.php
@@ -594,11 +594,14 @@ if ( ! class_exists( 'Give' ) ) :
 
 			require_once GIVE_PLUGIN_DIR . 'blocks/load.php';
 
-			//Include new reports page
-			require_once GIVE_PLUGIN_DIR . 'src/Reports/Reports.php';
-
 			//Include API
 			require_once GIVE_PLUGIN_DIR . 'src/API/API.php';
+
+			//Include Views
+			require_once GIVE_PLUGIN_DIR . 'src/Views/Views.php';
+
+			//$classes = get_declared_classes();
+			//print_r($classes);
 
 			if ( defined( 'WP_CLI' ) && WP_CLI ) {
 				require_once GIVE_PLUGIN_DIR . 'includes/class-give-cli-commands.php';
